% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abAnnotations.R
\name{ab_annotations}
\alias{ab_annotations}
\title{Get annotations from audioBlast}
\usage{
ab_annotations(...)
}
\arguments{
\item{...}{Fields and values to filter on. Any field defined by the Ann-o-mate project
may be filtered. The most common is likely to be taxon.}
}
\value{
A data frame of matching annotations
}
\description{
Search for annotated audio sections on audioBlast (via the ann-o-mate project).
}
\examples{
\dontrun{
ab_annotations(taxon="Gryllotalpa vineae")
}

}
