% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_seqss.R
\name{ab_seqss_nearestStart}
\alias{ab_seqss_nearestStart}
\title{Nearest start time}
\usage{
ab_seqss_nearestStart(...)
}
\arguments{
\item{...}{Fields and values to filter on.}
}
\value{
A data frame of matching annotations
}
\description{
Search audioBLAST! for recordings with a start time closest to
specified date/time which match specified criteria
}
\examples{
\dontrun{
ab_seqss_nearestStart(date="2020-05-15",time="1500")
}

}
