% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\docType{class}
\name{Annotation-class}
\alias{Annotation-class}
\title{A S4 class for annotations}
\description{
The \code{Annotation} class is used to store annotations on \code{Wave}-like objects.
}
\section{Slots}{

\describe{
\item{\code{file}}{File being annotated.}

\item{\code{metadata}}{A list for storing metadata.}

\item{\code{start}}{Start time of annotation.}

\item{\code{end}}{End time of annotation.}

\item{\code{low}}{Low frequency of annotation.}

\item{\code{high}}{High frequency of annotation.}

\item{\code{source}}{Source of annotation.}

\item{\code{type}}{Type of annotation.}

\item{\code{value}}{Value of annotation.}
}}

