% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\name{annotation}
\alias{annotation}
\title{Create a new Annotation object}
\usage{
annotation(
  file = NA_character_,
  metadata = list(),
  start = 0,
  end = Inf,
  low = 0,
  high = Inf,
  source = NA_character_,
  type = NA_character_,
  value = NA_character_
)
}
\arguments{
\item{file}{File being annotated.}

\item{metadata}{A list of metadata.}

\item{start}{Start time of annotation (seconds).}

\item{end}{End time of annotation (seconds).}

\item{low}{Low frequency of annotation (Hz).}

\item{high}{High frequency of annotation (Hz).}

\item{source}{Source of annotation.}

\item{type}{Type of annotation.}

\item{value}{Value of annotation.}
}
\value{
An Annotation object.
}
\description{
Create a new Annotation object
}
