% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\name{annotation_merge_overlapping}
\alias{annotation_merge_overlapping}
\title{Combine overlapping annotations}
\usage{
annotation_merge_overlapping(
  annotation1,
  annotation2,
  domain = "time",
  same.source = TRUE
)
}
\arguments{
\item{annotation1}{An Annotation object.}

\item{annotation2}{An Annotation object.}

\item{domain}{Domain of the annotations, either "time", "frequency", or "both".}

\item{same.source}{If TRUE, annotations must have the same source to be merged.}
}
\value{
A new Annotation object.
}
\description{
Combine overlapping annotations
}
