% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\name{AnnotationList2DataFrame}
\alias{AnnotationList2DataFrame}
\title{Convert a list of Annotation objects to a data.frame}
\usage{
AnnotationList2DataFrame(x)
}
\arguments{
\item{x}{A list of Annotation objects.}
}
\value{
A data.frame.
}
\description{
Converts a list of Annotation objects to a data.frame, with one column for
each metadata name in the list.
}
