% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\name{annotations_merge}
\alias{annotations_merge}
\title{Combine annotations}
\usage{
annotations_merge(annotations, domain = "time", same.source = TRUE)
}
\arguments{
\item{annotations}{A list of Annotation objects.}

\item{domain}{Domain of the annotations, either "time", "frequency", or "both".}

\item{same.source}{If TRUE, annotations must have the same source to be merged.}
}
\value{
A list of Annotation objects.
}
\description{
Checks a list of annotations for those that are overlapping, and returns a list
where overlapping annotations are merged.
}
