% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audioblast.R
\name{audioblast}
\alias{audioblast}
\title{Get data or analyses from audioBlast}
\usage{
audioblast(type, name, ...)
}
\arguments{
\item{type}{One of data, analysis, standalone.}

\item{name}{of data or analysis source.}

\item{...}{Fields and values to filter on. Any field defined by audioBLAST.}
}
\value{
A data frame
}
\description{
Search for recordings or analyses on audioBlast.
}
\examples{
\dontrun{
audioblast("data", "recordings", taxon="Gryllotalpa vineae")
}

}
