% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abAnnotations.R
\name{audioblast_annotations}
\alias{audioblast_annotations}
\title{Get annotations from audioBlast}
\usage{
audioblast_annotations(...)
}
\arguments{
\item{...}{Fields and values to filter on.}
}
\value{
A data frame of matching annotations
}
\description{
Search for annotated audio sections on audioBlast (via the ann-o-mate project).
}
\examples{
audioblast_annotations(taxon="Gryllotalpa vineae")
}
