% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audioblast.R
\name{audioblast_ASITSN}
\alias{audioblast_ASITSN}
\title{audioBlast - a stitch in time saves nine}
\usage{
audioblast_ASITSN(type, name, endpoint = NULL, ...)
}
\arguments{
\item{type}{One of data, analysis, standalone.}

\item{name}{Name of data or analysis source.}

\item{endpoint}{Optionally specify endpoint of an audioBlast module.}

\item{...}{Fields and values to filter on. Any field defined by audioBLAST.}
}
\description{
Sanity check before sending request to audioBlast
}
