% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audioblast.R
\name{audioblastDownload}
\alias{audioblastDownload}
\title{Download audio files from audioBlast}
\usage{
audioblastDownload(
  d,
  metadata = TRUE,
  skip.existing = TRUE,
  dir = ".",
  quiet = FALSE,
  on.issue = .audioblastIssue
)
}
\arguments{
\item{d}{Data returned from a search using audioBlast().}

\item{metadata}{If true saves the data in d as a csv file.}

\item{skip.existing}{If true will not overwrite existing files.}

\item{dir}{Directory to save files to.}

\item{quiet}{If true will not print progress.}

\item{on.issue}{Function to call on error or warning. By default \code{stop} to raise
a standard R error. Setting to \code{warning} will instead a warning.}
}
\value{
No return value, called for its side effect of downloading files.
}
\description{
Downloads audio files associated with a search using the audioBlast() function.
}
