% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audiomoth.R
\name{audiomoth_wave}
\alias{audiomoth_wave}
\title{Read AudioMoth metadata from a wave file}
\usage{
audiomoth_wave(filename)
}
\arguments{
\item{filename}{Path to the wave file to read}
}
\value{
A list of extracted parameters
}
\description{
Reads and parses metadata stored in wave files produced by
AudioMoth devices.
}
\examples{
\dontrun{
audiomoth_wavew("./FILENAME.WAV")
}

}
