% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audiomoth.R
\name{audiomothWave}
\alias{audiomothWave}
\title{Read AudioMoth metadata from a wave file}
\usage{
audiomothWave(filename)
}
\arguments{
\item{filename}{Path to the wave file to read}
}
\value{
A list of extracted parameters
}
\description{
Reads and parses metadata stored in wave files produced by
AudioMoth devices.
}
\examples{
\dontrun{
audiomothWave("./FILENAME.WAV")
}

}
