% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnetReticulate.R
\name{birdNetInstall}
\alias{birdNetInstall}
\title{Install the BirdNET environment}
\usage{
birdNetInstall(unattended = FALSE)
}
\arguments{
\item{unattended}{If TRUE then the function will not prompt the user to install
the environment in a non-interactive session.}
}
\description{
This function installs BirdNET in the \code{ssd_birdnet} environment using \code{reticulate}.
}
\examples{
\dontrun{
birdnetInstall()
birdNetInstall(unattended=TRUE)
}
}
