% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allChannels.R
\name{channels_se}
\alias{channels_se}
\title{Channels for sound ecology}
\usage{
channels_se(...)
}
\arguments{
\item{...}{Export from a bioacoustic index function from the soundecology package}
}
\value{
A list containing the value calculated for the left channel.
}
\description{
Used to process the output of acoustic index functions from the soundecology
package when using allChannels.
}
