% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radialPolygon.R
\name{circularise}
\alias{circularise}
\title{Circularise a dataset}
\usage{
circularise(values)
}
\arguments{
\item{values}{A vector if values}
}
\value{
The input vector with its first value appended, so that it closes on itself.
}
\description{
When plotting rings or horizons that are meant to cover the entirety of the time period in a
dielPlot() or yearlyPlot() this function append the beginning values to the end to ensure an entire
loop is created.
}
