% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertAngles.R
\name{convert2degrees}
\alias{convert2degrees}
\title{Convert angle to degrees}
\usage{
convert2degrees(A, input = "radians")
}
\arguments{
\item{A}{The angle value to convert}

\item{input}{The unit of angle to convert, allowed values are "radians".}
}
\value{
The numeric value in degrees
}
\description{
Converts angle measurements into degrees
}
