% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTemperatures.R
\name{convert2Fahrenheit}
\alias{convert2Fahrenheit}
\title{Convert temperature to Fahrenheit}
\usage{
convert2Fahrenheit(temp, input)
}
\arguments{
\item{temp}{The value of the temperature to convert}

\item{input}{The unit of the temperature to convert, allowed values are "K", "C".}
}
\value{
No return value, this conversion is not implemented and the function raises an error.
}
\description{
Converts temperature measurements into Fahrenheit
}
\examples{
\dontrun{
convert2Fahrenheit(15, input = "C")
}

}
