% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTemperatures.R
\name{convert2Kelvin}
\alias{convert2Kelvin}
\title{Convert temperature to Kelvin}
\usage{
convert2Kelvin(temp, input = "C")
}
\arguments{
\item{temp}{The value of the temperature to convert}

\item{input}{The unit of the temperature to convert, allowed values are "C", "F".}
}
\value{
Numeric value in Kelvin
}
\description{
Converts temperature measurements into Kelvin
}
\examples{
convert2Kelvin(15, input="C")
convert2Kelvin(15, input="F")

}
