% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertAngles.R
\name{convert2radians}
\alias{convert2radians}
\title{Convert angle to radians}
\usage{
convert2radians(A, input = "degrees")
}
\arguments{
\item{A}{The angle value to convert}

\item{input}{The unit of angle to convert, allowed values are "degrees".}
}
\value{
The numeric value in radians
}
\description{
Converts angle measurements into radians
}
