% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTime.R
\name{convert2seconds}
\alias{convert2seconds}
\title{Convert time to seconds}
\usage{
convert2seconds(T, input = "minutes", origin = "day")
}
\arguments{
\item{T}{The time value to convert}

\item{input}{The unit of time to convert, allowed values are "minutes",
"hours", "days", "years", "HHMM".}

\item{origin}{For POSIX whether to return relative to start of day ("day") or Unix epoch ("unix")}
}
\value{
The numeric value in seconds
}
\description{
Converts time measurements into seconds
}
