% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2Wave.R
\name{data2Wave}
\alias{data2Wave}
\title{Convert data into a Wave object}
\usage{
data2Wave(
  left,
  samp.rate = 44100,
  bit = 16,
  unit = NULL,
  remove.offset = TRUE,
  normalise = TRUE
)
}
\arguments{
\item{left}{Data for mono audio channel}

\item{samp.rate}{Sampling rate for Wave object}

\item{bit}{Bit depth of Wave object}

\item{unit}{See tuneR::normalize. If NULL this is handled automatically.}

\item{remove.offset}{If TRUE any DC offset is removed}

\item{normalise}{If TRUE the output Wave is normalised to the range implied by
\code{unit}, which for the default of NULL is the range of the bit depth rather
than -1:1.}
}
\value{
A mono Wave object.
}
\description{
Make a sequence of data into a normalised Wave object.
}
\examples{
pattern <- seq(from=-1, to=1, length.out=100)
data <- rep.int(pattern, 100)
w <- data2Wave(data)
}
