% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dayPhase.R
\name{dayPhases}
\alias{dayPhases}
\title{Phases of day}
\usage{
dayPhases(time = as.Date(Sys.time()), lat = 50.1, lon = 1.83, tz = "UTC")
}
\arguments{
\item{time}{A time object representing the start time of a recording}

\item{lat}{Latitude of recording device}

\item{lon}{Longitude of recording device}

\item{tz}{Time-zone of recording device when recording was made}
}
\value{
A matrix with a row for each phase of the day and columns giving its start and end.
}
\description{
Wrapper for suncalc::getSunlightTimes that formats output for this package.
}
