% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dayPhase.R
\name{daysPhases}
\alias{daysPhases}
\title{Phases of days}
\usage{
daysPhases(
  date = Sys.Date(),
  period = "year",
  plot = FALSE,
  lat = 50.1,
  lon = 1.83,
  tz = "UTC"
)
}
\arguments{
\item{date}{A time object representing the start time of a recording}

\item{period}{"month" or "year"}

\item{plot}{If true plots the data, default FALSE}

\item{lat}{Latitude of recording device}

\item{lon}{Longitude of recording device}

\item{tz}{Time-zone of recording device when recording was made}
}
\value{
A data frame with a row for each day, giving the times of each phase of the day and the state of the moon.
}
\description{
Phases of days
}
