% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultCluster.R
\name{defaultCluster}
\alias{defaultCluster}
\title{Create Default Cluster for Windowing}
\usage{
defaultCluster(fork = TRUE)
}
\arguments{
\item{fork}{If TRUE uses forking to create the cluster (Unix like systems only)}
}
\value{
A cluster object for parallel processing
}
\description{
Creates a default cluster using one less than the total cores available on
the system. By default this uses forking, which is not be available on Windows.
Hence, the fork parameter has no effect on Windows.
}
\examples{
\dontrun{
cl <- defaultCluster()
stopCluster(cl)
cl <- defaultCluster(FALSE)
stopCluster(cl)
}

}
