% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dielPlot.R
\name{dielFraction}
\alias{dielFraction}
\title{Calculate the fraction of a day given by a value}
\usage{
dielFraction(t, input = "POSIX", unit = "radians")
}
\arguments{
\item{t}{Object to be converted to a fraction}

\item{input}{One of POSIX (default) or HHMM}

\item{unit}{If set to radians outputs a position around a circle. If set to fraction outputs the raw fraction.}
}
\value{
The position of the time within the day, in radians or as a fraction of a day.
}
\description{
Given an object that can be coerced to POSIXlt or is in a supported string
format, return the fraction of a day represented by the object.
}
