% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaveFilter.R
\name{doNowt}
\alias{doNowt}
\title{Do nothing}
\usage{
doNowt(x)
}
\arguments{
\item{x}{A thing.}
}
\value{
The same thing.
}
\description{
This function does nothing to the Wave (or any object).
}
\keyword{internal}
