% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\name{.annotation_can_merge}
\alias{.annotation_can_merge}
\title{Check if two annotations can be merged}
\usage{
.annotation_can_merge(annotation1, annotation2, same.source = TRUE)
}
\arguments{
\item{annotation1}{An Annotation object.}

\item{annotation2}{An Annotation object.}

\item{same.source}{If TRUE, annotations must have the same source to be merged.}
}
\value{
TRUE if the annotations can be merged, FALSE otherwise.
}
\description{
Check if two annotations can be merged
}
