% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\name{.annotation_check_overlap}
\alias{.annotation_check_overlap}
\title{Check if two annotations overlap or are continuous}
\usage{
.annotation_check_overlap(annotation1, annotation2, domain = "time")
}
\arguments{
\item{annotation1}{An Annotation object.}

\item{annotation2}{An Annotation object.}

\item{domain}{Domain of the annotations, either "time", "frequency", or "both".}
}
\value{
TRUE if the annotations overlap, FALSE otherwise.
}
\description{
Check if two annotations overlap or are continuous
}
