% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\name{.merge_annotations}
\alias{.merge_annotations}
\title{Combine annotations helper function}
\usage{
.merge_annotations(annotations, domain = "time")
}
\arguments{
\item{annotations}{A list of Annotation objects.}

\item{domain}{Domain of the annotations, either "time" or "frequency".}
}
\value{
A list of Annotation objects.
}
\description{
Checks a list of annotations for those that are overlapping in the time or
frequency domain, and returns a list where overlapping annotations are merged.
}
\details{
The exported function \code{merge_annotations()} handles sanity checks and calls this function.
}
