% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scikit-maad-classes.R
\name{.spectrogram_maad_power}
\alias{.spectrogram_maad_power}
\title{Helper function to convert a Wave-like object to scikit-maad spectrogram}
\usage{
.spectrogram_maad_power(object)
}
\arguments{
\item{object}{A Wave or WaveMC object (or Tagged equivalent),
or a spectrogram_maad object}
}
\value{
A spectrogram_maad object
}
\description{
Helper function to convert a Wave-like object to scikit-maad spectrogram
}
