% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scikit-maad-classes.R
\name{.spectrum_maad}
\alias{.spectrum_maad}
\title{Helper function to convert a Wave-like object to scikit-maad spectrum}
\usage{
.spectrum_maad(object)
}
\arguments{
\item{object}{A Wave or WaveMC object (or Tagged equivalent),
or a spectrum}
}
\value{
A spectrum list
}
\description{
Helper function to convert a Wave-like object to scikit-maad spectrum
}
