% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dielPlot.R
\name{emptyDiel}
\alias{emptyDiel}
\title{Create an empty diel plot}
\usage{
emptyDiel(method = "plotrix", rot = pi)
}
\arguments{
\item{method}{Plotting package to use}

\item{rot}{Rotation of the origin (defaults to pi)}
}
\value{
Called for its side effect of drawing a plot. The return value is that of the underlying plotting function and should not be relied on.
}
\description{
Create a diel plot with labels but without sun altitude or times of day plotted.
}
