% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearlyPlot.R
\name{emptyYearly}
\alias{emptyYearly}
\title{Create an empty yearly plot}
\usage{
emptyYearly(year = 2022, method = "plotrix", rot = pi)
}
\arguments{
\item{year}{Year to plot (allows for leap years)}

\item{method}{Plotting package to use}

\item{rot}{Rotation of the origin (defaults to pi)}
}
\value{
Called for its side effect of drawing a plot. The return value is that of the underlying plotting function and should not be relied on.
}
\description{
Create a yearly plot with labels but without sun or night duration plotted.
}
