% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaveFilter.R
\name{filterw}
\alias{filterw}
\title{Apply a WaveFilter object to a Wave object}
\usage{
filterw(w, filt)
}
\arguments{
\item{w}{A Wave object.}

\item{filt}{Wave object with the selected filter applied.}
}
\description{
A WaveFilter object is an object containing information necessary for the
filterw function to apply the filter to a Wave object. This is designed to
allow a pipe operator (either magrittr or base R) to be used to apply filters
to a Wave in a pipeline.
}
\details{
Supported filters include those from the seewave package.
}
