% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencyStats.R
\name{frequencyStats}
\alias{frequencyStats}
\title{Various measurements of frequency values for a Wave object}
\usage{
frequencyStats(wave, wave_spec = NULL, warn = TRUE, lowcut = 1, plot = FALSE)
}
\arguments{
\item{wave}{A Wave object. Stereo Wave and WaveMC objects are passed to
allChannels(), which returns one set of statistics per channel.}

\item{wave_spec}{A precomputed spectrum (optional, if not present will be generated)}

\item{warn}{If TRUE provides warnings when values are not consistent}

\item{lowcut}{Frequency (in kHz) values below which are ignored.}

\item{plot}{IF TRUE displays values}
}
\value{
A list of frequency statistics, with an entry for each of the -3dB and -10dB thresholds.
}
\description{
Calculates the peak, centre, bandwidth and quality factor. The quality factor (Q) is calculated at both
-3dB and -10dB as discussed by Bennett-Clark (1999)  <doi: 10.1080/09524622.1999.9753408>.
}
