% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{frequencyWS}
\alias{frequencyWS}
\title{Get the frequency from wavelength and speed of sound}
\usage{
frequencyWS(wl, s)
}
\arguments{
\item{wl}{Wavelength}

\item{s}{Speed of sound}
}
\description{
TODO: Description
}
