% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateNoise.R
\name{generateNoise}
\alias{generateNoise}
\title{Windowing Function for Wave Objects}
\usage{
generateNoise(wave, noise = "white", noiseAdd = FALSE,
  noiseRatio = 0.5, output = "file", plot = FALSE)
}
\arguments{
\item{wave}{A Wave object}

\item{window.length}{The lag used to create the A-matrix}

\item{window.overlap}{A matrix used to code the Duration-Shape pairs}

\item{max_D}{The maximum Duration to code}

\item{FUN}{If TRUE plots the workings of the coding algorithm}

\item{...}{Additional parameters to FUN}

\item{cluster}{A cluser form the 'parallel' package for multicore computation}
}
\description{
Seperates a Wave object into windows of a defined length and runs a function on the window section. Windows may overlap, and the function can make use of 'parallel' package for multicore processing.
}
\examples{
library(tuneR)
wave <- readWave(system.file("extdata", "1.wav", package="tdsc"))
t <- tdsc(wave)
}
\keyword{wave}
