% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateTimeMask.R
\name{generateTimeMasked}
\alias{generateTimeMasked}
\title{Generate time masked Wave-like objects}
\usage{
generateTimeMasked(wave, method = "squarewave")
}
\arguments{
\item{wave}{A \code{Wave}-like object (or a list of \code{Wave}-like objects).}

\item{method}{The method to use for time masking (one of "squarewave").}
}
\description{
Given a \code{Wave}-like object (or a list of \code{Wave}-like objects), generate
new \code{Wave}-like objects with time masking.
}
