% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scikit-maad.R
\name{getMaad}
\alias{getMaad}
\title{Import scikit-maad}
\usage{
getMaad()
}
\value{
The scikit-maad Python module, with its sound and features submodules attached.
}
\description{
Imports the scikit-maad object into R. This allows reuse of the same object between
function calls.
}
