% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleSpeech.R
\name{gs_transcribe}
\alias{gs_transcribe}
\title{Google Speech API Transcribe}
\usage{
gs_transcribe(filename, bucket = NULL, ...)
}
\arguments{
\item{filename}{Path to file for analysis}

\item{bucket}{Storage bucket on Google Cloud for larger files}

\item{...}{Additional arguments to pass to gl_speech()}
}
\description{
Wrapper around various Google packages to simplify speech transcription.
}
