% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{labelPadding}
\alias{labelPadding}
\title{Pad labels with interval}
\usage{
labelPadding(t, pad = 0.5, max_t = NULL)
}
\arguments{
\item{t}{Transcript from Google Speech API}

\item{pad}{Amount of time (in seconds) to add to start and end}

\item{max_t}{Optional. The duration of the file, so padding does not exceed length of file.}
}
\value{
A modified Google Speech API transcript object
}
\description{
Takes labels from Google Speech API transcript and pads the time by a specified number of seconds.
}
\examples{
\dontrun{
labelPadding(t, pad=2, max_t=duration(wave))
}

}
