% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scikit-maad-indices.R
\name{maad_acoustic_complexity_index}
\alias{maad_acoustic_complexity_index}
\title{Compute the acoustic complexity index using scikit-maad}
\usage{
maad_acoustic_complexity_index(object, maad = NULL)
}
\arguments{
\item{object}{A Wave object or a spectrogram_maad object generated by
\code{\link{maad_spectrogram}}. If a Wave-like object is provided, the
spectrogram will be calculated using the default parameters.}

\item{maad}{An optional maad object. If not provided, one will be created using \code{getMaad()}.}
}
\value{
List comprising:
\item{ACI_xx}{Acoustic Complexity Index.}
\item{ACI_per_bin}{Acoustic Complexity Index.}
\item{ACI_sum}{Sum of ACI value per frequency bin (Common definition)}
}
\description{
ACI depends on the duration of the spectrogram as the derivation of the signal
is normalized by the sum of the signal. Thus, if the background noise is high
due to high acoustic activity the normalization by the sum of the signal
reduced ACI. So ACI is low when there is no acoustic activity or high
acoustic activity with continuous background noise. ACI is high only when
acoustic activity is medium, with sounds well above the background noise.
}
\details{
For addition documentation see
\url{https://scikit-maad.github.io/generated/maad.features.temporal_acoustic_complexity_index.html}
\insertCite{pieretti2011}{sonicscrewdriver}.
}
\references{
\insertAllCited{}
}
