% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scikit-maad-indices.R
\name{maad_frequency_entropy}
\alias{maad_frequency_entropy}
\title{Compute the frequency entropy using scikit-maad}
\usage{
maad_frequency_entropy(object, compatibility = "QUT", maad = NULL)
}
\arguments{
\item{object}{A Wave object or a spectrogram_maad object generated by
\code{\link{maad_spectrogram}}. If a Wave-like object is provided, the
spectrogram will be calculated using the default parameters.}

\item{compatibility}{One of "QUT" \insertCite{qut}{sonicscrewdriver} or "seewave"
\insertCite{seewave2008}{sonicscrewdriver}. Default is "QUT".}

\item{maad}{An optional maad object. If not provided, one will be created using \code{getMaad()}.}
}
\value{
List comprising:
\item{Hf}{Frequency entropy.}
\item{Ht_per_bin}{Temporal entropy per frequency bin.}
}
\description{
The spectral entropy of a signal measures the energy dispersion along
frequencies. Low values indicates a concentration of energy around a narrow
frequency band. If the DC value is not removed before processing the large
peak at f=0Hz will lower the entropy of the signal.
}
\details{
For addition documentation see
\url{https://scikit-maad.github.io/generated/maad.features.frequency_entropy.html}
}
\references{
\insertAllCited{}
}
