% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scikit-maad-indices.R
\name{maad_number_of_peaks}
\alias{maad_number_of_peaks}
\title{Compute the number of peaks using scikit-maad}
\usage{
maad_number_of_peaks(object, ..., maad = NULL)
}
\arguments{
\item{object}{A Wave object or a spectrogram_maad object generated by
\code{\link{maad_spectrogram}}. If a Wave-like object is provided, the
spectrogram will be calculated using the default parameters.}

\item{...}{Additional arguments passed to \code{maad.features.number_of_peaks}.}

\item{maad}{An optional maad object. If not provided, one will be created using \code{getMaad()}.}
}
\value{
Numeric number of peaks
}
\description{
Count the number of frequency peaks on a mean spectrum. This function was
adapted from the function fpeaks of the R package Seewave
\insertCite{seewave2008}{sonicscrewdriver}.
}
\details{
For addition documentation see
\url{https://scikit-maad.github.io/generated/maad.features.number_of_peaks.html}
}
