% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scikit-maad-indices.R
\name{maad_spectral_activity}
\alias{maad_spectral_activity}
\title{Compute the spectral activity using scikit-maad}
\usage{
maad_spectral_activity(object, dB_threshold = 6, maad = NULL)
}
\arguments{
\item{object}{A Wave object or a spectrogram_maad object.}

\item{dB_threshold}{dB threshold of activity (default = 6).}

\item{maad}{An optional maad object. If not provided, one will be created using \code{getMaad()}.}
}
\value{
The function computes for each frequency bin:
\item{ACTfract}{Proportion (fraction) of points above the threshold.}
\item{ACTcount}{Total number of points above the threshold.}
\item{ACTmean}{Mean value (in dB) of the portion of the signal above the threshold.}
}
\description{
Acoustic activity corresponds to the portion of the spectrogram above a
threshold frequency per frequency along time axis \insertCite{towsey2017}{sonicscrewdriver}.
}
\references{
\insertAllCited{}
}
