% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scikit-maad-indices.R
\name{maad_spectral_entropy}
\alias{maad_spectral_entropy}
\title{Compute the frequency entropy using scikit-maad}
\usage{
maad_spectral_entropy(object, flim = NULL, maad = NULL)
}
\arguments{
\item{object}{A Wave object or a spectrogram_maad object generated by
\code{\link{maad_spectrogram}}. If a Wave-like object is provided, the
spectrogram will be calculated using the default parameters.}

\item{flim}{A numeric vector of length 2 specifying the frequency limits for the entropy calculation.}

\item{maad}{An optional maad object. If not provided, one will be created using \code{getMaad()}.}
}
\value{
List comprising:
\item{EAS}{Entropy of Average Spectrum.}
\item{ECU}{Entropy of spectral variance (along the time axis for each frequency).}
\item{ECV}{Entropy of Coefficient of Variation (along the time axis for each frequency).}
\item{EPS}{Entropy of spectral maxima (peaks).}
\item{EPS_KURT}{Kurtosis of spectral maxima.}
\item{EPS_SKEW}{Skewness of spectral maxima.}
}
\description{
Compute different entropies based on the average spectrum, its variance, and its maxima.
}
\details{
For addition documentation see
\url{https://scikit-maad.github.io/generated/maad.features.spectral_entropy.html}
}
