% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scikit-maad-sound.R
\name{maad_spectrogram}
\alias{maad_spectrogram}
\title{Compute a spectrogram of a sound wave using scikit-maad}
\usage{
maad_spectrogram(wave, mode = "power", ..., maad = NULL)
}
\arguments{
\item{wave}{A Wave object}

\item{mode}{The type of spectrogram to compute. Options are "power", "amplitude" "complex".
Default is "power".}

\item{...}{Additional arguments to pass to \code{maad.sound.spectrogram}.}

\item{maad}{An optional \code{maad} object. If not provided, one will be created using \code{getMaad()}.}
}
\value{
Generically a \code{spectrogram_maad} object.
}
\description{
This is a wrapper function for the \code{maad.sound.spectrogram} function
from the scikit-maad package for Python. It computes the spectrogram of a
sound wave. Further usage details are provided at
\url{https://maad.readthedocs.io/en/latest/maad.sound.html#maad.sound.spectrogram}.
}
