% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scikit-maad-sound.R
\name{maad_spectrum}
\alias{maad_spectrum}
\title{Compute a spectrum of a sound wave using scikit-maad}
\usage{
maad_spectrum(wave, ..., maad = NULL)
}
\arguments{
\item{wave}{A Wave object}

\item{...}{Additional arguments to pass to \code{maad.sound.spectrum}.}

\item{maad}{An optional \code{maad} object. If not provided, one will be created using \code{getMaad()}.}
}
\value{
A list comprising:
\item{pxx}{Power spectral density estimate.}
\item{f_idx}{Index of sample frequencies.}
}
\description{
This is a wrapper function for the \code{maad.sound.spectrum} function.
It computes the spectrum of a sound wave. Further usage details are provided at
\url{https://maad.readthedocs.io/en/latest/maad.sound.html#maad.sound.spectrum}.
}
