% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scikit-maad-indices.R
\name{maad_temporal_activity}
\alias{maad_temporal_activity}
\title{Compute the acoustic activity index using scikit-maad}
\usage{
maad_temporal_activity(
  wave,
  dB_threshold = 3,
  mode = "fast",
  Nt = 512,
  maad = NULL
)
}
\arguments{
\item{wave}{A Wave object.}

\item{dB_threshold}{dB threshold of activity (default = 3).}

\item{mode}{Mode of the envelope calculation. Can be "fast" or "Hilbert".}

\item{Nt}{Size of each frame. The largest, the highest is the approximation.}

\item{maad}{An optional maad object. If not provided, one will be created using \code{getMaad()}.}
}
\value{
List of calculated values, comprising:
\item{ACTfrac}{fraction)of points above the threshold for each frequency bin.}
\item{ACTcount}{total number of points above the threshold for each frequency bin.}
\item{ACTmean}{mean value (in dB) of the portion of the signal above the threshold.}
}
\description{
For addition documentation see
\url{https://scikit-maad.github.io/generated/maad.features.temporal_activity.html}
\insertCite{towsey2013}{sonicscrewdriver}.
}
\references{
\insertAllCited{}
}
