% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scikit-maad-indices.R
\name{maad_temporal_entropy}
\alias{maad_temporal_entropy}
\title{Compute the temporal entropy using scikit-maad}
\usage{
maad_temporal_entropy(
  wave,
  compatibility = "QUT",
  mode = "fast",
  Nt = 512,
  maad = NULL
)
}
\arguments{
\item{wave}{A Wave object.}

\item{compatibility}{One of "QUT" \insertCite{qut}{sonicscrewdriver}, "seewave" \insertCite{seewave2008}{sonicscrewdriver}.}

\item{mode}{Mode of the envelope calculation. Can be "fast" or "hilbert".}

\item{Nt}{Size of each frame. The largest, the highest is the approximation.}

\item{maad}{An optional maad object. If not provided, one will be created using \code{getMaad()}.}
}
\value{
Numeric entropy of the envelope.
}
\description{
For addition documentation see
\url{https://scikit-maad.github.io/generated/maad.features.temporal_entropy.html}.
}
\references{
\insertAllCited{}
}
