% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scikit-maad-indices.R
\name{maad_temporal_events}
\alias{maad_temporal_events}
\title{Compute the acoustic event index using scikit-maad}
\usage{
maad_temporal_events(
  wave,
  dB_threshold = 3,
  rejectDuration = 0.1,
  mode = "fast",
  Nt = 512,
  maad = NULL
)
}
\arguments{
\item{wave}{A Wave object.}

\item{dB_threshold}{dB threshold of activity (default = 3).}

\item{rejectDuration}{Duration of the silence to reject an acoustic event (default = 0.1).}

\item{mode}{Mode of the envelope calculation. Can be "fast" or "Hilbert".}

\item{Nt}{Size of each frame. The larger, the highest is the approximation.}

\item{maad}{An optional maad object. If not provided, one will be created using \code{getMaad()}.}
}
\value{
List of calculated values, comprising:
\item{EVTfrac}{fraction of points above the threshold for each frequency bin.}
\item{EVTcount}{total number of points above the threshold for each frequency bin.}
\item{EVTmean}{mean value (in dB) of the portion of the signal above the threshold.}
\item{EVN}{logical vector or matrix with 1 corresponding to event position.}
}
\description{
An acoustic event corresponds to the period of the signal above a threshold.
An acoustic event could be short (at list one point if rejectDuration is
None) or very long (the duration of the  entire audio). Two acoustic events
are separated by a period with low audio signal (i.e. below the threshold).
}
\details{
For addition documentation see
\url{https://scikit-maad.github.io/generated/maad.features.temporal_events.html}
\insertCite{towsey2013}{sonicscrewdriver} \insertCite{qut}{sonicscrewdriver}.
}
\references{
\insertAllCited{}
}
