% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scikit-maad-indices.R
\name{maad_temporal_median}
\alias{maad_temporal_median}
\title{Compute the temporal envelope median using scikit-maad}
\usage{
maad_temporal_median(wave, mode = "fast", Nt = 512, maad = NULL)
}
\arguments{
\item{wave}{A Wave object.}

\item{mode}{Mode of the envelope calculation. Can be "fast" or "Hilbert".}

\item{Nt}{Size of each frame. The largest, the highest is the approximation.}

\item{maad}{An optional maad object. If not provided, one will be created using \code{getMaad()}.}
}
\value{
Numeric median of the envelope.
}
\description{
For addition documentation see
\url{https://scikit-maad.github.io/generated/maad.features.temporal_median.html}.
}
