% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scikit-maad.R
\name{maadInstall}
\alias{maadInstall}
\title{Install the scikit-maad environment}
\usage{
maadInstall(unattended = FALSE)
}
\arguments{
\item{unattended}{If TRUE then the function will not prompt the user to install
the environment in a non-interactive session.}
}
\description{
This function installs scikit-maad in the \code{ssd_scikit-maad} environment using \code{reticulate}.
}
\examples{
\dontrun{
maadInstall()
maadInstall(unattended=TRUE)
}
}
