% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{normalise}
\alias{normalise}
\title{Normalise a Wave object}
\usage{
normalise(wave, unit = NULL, ...)
}
\arguments{
\item{wave}{Wave or WaveMC object}

\item{unit}{If not null behaves as in normalize() from tuneR, if null the unit
is automatically identified.}

\item{...}{Additional arguments passed to normalize() from tuneR}
}
\value{
Normalised Wave or WaveMC object
}
\description{
Similar to normalize() from the tuneR package but automatically identifies the
unit parameter.
}
