% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntd.R
\name{ntd}
\alias{ntd}
\title{Natural Time Domain}
\usage{
ntd(wave, events, FUN, normalise = FALSE)
}
\arguments{
\item{wave}{A Wave object containing pulses}

\item{events}{Onset of detected events, e.g. from pulseDetction()}

\item{FUN}{The function to run}

\item{normalise}{If TRUE the output is a probability density}
}
\description{
Runs a function on the wave and outpiuts values in the Natural Time Domain.
}
