% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseFilename.R
\name{parseFilename}
\alias{parseFilename}
\title{Parse a filename}
\usage{
parseFilename(string)
}
\arguments{
\item{string}{A filename}
}
\description{
Attempts to extract menaingful information froma  filename.
}
