% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulseDetection_dietrich.R
\name{pd_dietrich2004}
\alias{pd_dietrich2004}
\title{Pulse detection using Dietrich (2004)}
\usage{
pd_dietrich2004(
  wave,
  U = 120,
  gamma = 0.05,
  alpha = 1.4,
  scaling = 32,
  V = 480,
  psi = 1
)
}
\arguments{
\item{wave}{A Wave object}

\item{U}{Window length}

\item{gamma}{Gamma}

\item{alpha}{Alpha}

\item{scaling}{Scaling}

\item{V}{V Window length}

\item{psi}{Psi}
}
\value{
A list of input values plus the onset and offset times of pulses
}
\description{
Detects pulses in a Wave using the method described in Dietrich et al (2004) \url{doi:10.1016/j.patcog.2004.04.004}.
}
