% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulseDetection_simple.R
\name{pd_simple}
\alias{pd_simple}
\title{Simplified pulse detection using Dietrich (2004)}
\usage{
pd_simple(
  wave,
  U = 120,
  gamma = 0.05,
  alpha = 1.4,
  scaling = 32,
  V = 480,
  psi = 1
)
}
\arguments{
\item{wave}{A Wave object}

\item{U}{Window length}

\item{gamma}{Gamma}

\item{alpha}{Alpha}

\item{scaling}{Scaling}

\item{V}{V Window length}

\item{psi}{Psi}
}
\value{
A list of input values plus the onset and offset times of pulses.
}
\description{
Detects pulses in a Wave.
}
