% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcis.R
\name{plot.fcis}
\alias{plot.fcis}
\title{Plot a false-colour index spectrogram}
\usage{
\method{plot}{fcis}(
  x,
  xlab = NULL,
  ylab = "Frequency (kHz)",
  main = "False-colour index spectrogram",
  ...
)
}
\arguments{
\item{x}{An object of class "fcis"}

\item{xlab}{Label for the x axis}

\item{ylab}{Label for the y axis}

\item{main}{Title for the plot}

\item{...}{Additional parameters passed to plot()}
}
\value{
No return value, called for its side effect of drawing a plot.
}
\description{
Plot a false-colour index spectrogram
}
\examples{
\dontrun{
plot(fcis(wave, window.length=44100))
}
}
