% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFileSizes.R
\name{printHumanBytes}
\alias{printHumanBytes}
\title{Prints bytes in human readable form}
\usage{
printHumanBytes(S)
}
\arguments{
\item{S}{Number of bytes}
}
\description{
Given an input of bytes calculates the result in a sensible output unit (e.g.
MB, GB, PB) and prints the output.
}
