% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseudoWave.R
\docType{methods}
\name{-,PseudoWave,numeric-method}
\alias{-,PseudoWave,numeric-method}
\title{PseudoWave scalar subtraction}
\usage{
\S4method{-}{PseudoWave,numeric}(e1, e2)
}
\arguments{
\item{e1}{Input 1}

\item{e2}{Input 2}
}
\value{
A PseudoWave object with e2 subtracted from its offset.
}
\description{
PseudoWave scalar subtraction
}
