% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulse.R
\name{pulse}
\alias{pulse}
\title{Generate a single pulse}
\usage{
pulse(
  type = "dirac",
  leading = 22050,
  pulse.length = 1,
  duration = samp.rate,
  samp.rate = 44100,
  bit = 1,
  pcm = FALSE,
  stereo = FALSE,
  output = "Wave",
  invert = FALSE
)
}
\arguments{
\item{type}{Either "dirac" or "square".}

\item{leading}{The number of samples before the pulse.}

\item{pulse.length}{The number of samples in the pulse (for "square").}

\item{duration}{The total number of samples generated.}

\item{samp.rate}{The sample rate.}

\item{bit}{The bit depth.}

\item{pcm}{Whether Wave generated is PCM (see tuneR).}

\item{stereo}{Whether Wave generated is stereo.}

\item{output}{The output format ("Wave").}

\item{invert}{Whether to invert the pulse.}
}
\value{
Specified by output.
}
\description{
Generate a single pulse, either a Dirac pulse (Dirac delta)  or a square pulse.
}
