% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulseIntervals.R
\name{pulseIntervals}
\alias{pulseIntervals}
\title{Pulse intervals}
\usage{
pulseIntervals(pulses, nsd = 2)
}
\arguments{
\item{pulses}{The result of a pulseDetection.}

\item{nsd}{The number of standard deviations each sid of the mean pulse interval to discard}
}
\value{
A list of onset and offset times for pulses
}
\description{
Used to locate area of no pulses from the results of pulseDetection().
}
