% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echolocation.R
\name{radar.range}
\alias{radar.range}
\title{Radar range}
\usage{
radar.range(t, c = soundSpeedMedium(medium = "air"))
}
\arguments{
\item{t}{Time in seconds}

\item{c}{Speed of sound in transmission medium m/s (by default air)}
}
\value{
Distance to object
}
\description{
Calculates the distance of an object based on the round trip time of an
echolocation pulse
}
\examples{
radar.range(2)
radar.range(2, c=343)
radar.range(2, c=soundSpeedMedium("sea water"))

}
