% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audacity.R
\name{readAudacityLabels}
\alias{readAudacityLabels}
\title{Read an Audacity label file}
\usage{
readAudacityLabels(file, output = "annotations")
}
\arguments{
\item{file}{Path to the Audacity label file.}

\item{output}{One of "annotations" or "data.frame".}
}
\value{
A list of Annotation objects, or a data frame if output is "data.frame".
}
\description{
Reads an Audacity label file and returns either a list of \code{Annotation} objects
or a data frame.
}
