% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audioblast.R
\name{readAudioBlast}
\alias{readAudioBlast}
\title{Download recording from audioBlast as a PseudoWave}
\usage{
readAudioBlast(source, id)
}
\arguments{
\item{source}{source field from audioblast.}

\item{id}{id field from audioblast.}
}
\value{
A PseudoWave object
}
\description{
Downloads a single recording metadata from audioBlast as a PseudoWave object.
}
