% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdNet.R
\name{readBirdNet}
\alias{readBirdNet}
\title{Read output files from BirdNet Analyser}
\usage{
readBirdNet(file, filename_parsing = "none")
}
\arguments{
\item{file}{Filename or directory}

\item{filename_parsing}{Allows for filename parsing, accepted values are one of none, audiomoth, timestamp.}
}
\value{
A data frame.
}
\description{
Reads a single file, or directory of files, output by BirdNet Analyser.
}
