% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respeaker6.R
\name{readRespeaker6}
\alias{readRespeaker6}
\title{Read a file from Seeed Studio Respeaker 6 mic array}
\usage{
readRespeaker6(filename, from = 1, to = Inf, units = "samples", header = FALSE)
}
\arguments{
\item{filename}{file to read}

\item{from}{See readWave}

\item{to}{See readWave}

\item{units}{See readWave}

\item{header}{See readWave}
}
\value{
Missing value of n or t
}
\description{
The Seeed Studio Respeaker-6 when used as described in the documentation saves an eight
channel audio file with channels 7 and 8 not containing input audio. This function reads
such a file and saves it as a six channel file.
}
