% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeRegion.R
\name{region}
\alias{region}
\title{Specify a region with a file to analyse}
\usage{
region(unit, from = 0, to = Inf)
}
\arguments{
\item{unit}{Unit of time (one of samples, seconds, minutes, hours)}

\item{from}{Start time}

\item{to}{End time}
}
\value{
A TimeRegion object.
}
\description{
Specifies a time-bounded region to analyse.
}
