% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheepFrequencyStats.R
\docType{data}
\name{sheepFrequencyStats}
\alias{sheepFrequencyStats}
\title{Sheep frequencyStats}
\format{An object of class \code{list} of length 3.}
\usage{
sheepFrequencyStats
}
\description{
The frequencyStats of the sheep data file from the seewave package.
}
\keyword{datasets}
