% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shimmer.R
\name{shimmer}
\alias{shimmer}
\title{Calculate the shimmer in a Wave object}
\usage{
shimmer(wave)
}
\arguments{
\item{wave}{A Wave object}
}
\value{
The shimmer in decibels
}
\description{
Shimmer is a measure of the variability of amplitude between successive periods
in the waveform. It is returned in decibels, as the mean absolute difference in
level between the peak amplitudes of consecutive periods. A waveform of constant
amplitude has a shimmer of zero.
}
\examples{
\dontrun{
shimmer(sheep)
}
}
