% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\name{sort_annotations}
\alias{sort_annotations}
\title{Sort annotations}
\usage{
sort_annotations(annotations, domain = "time", decreasing = FALSE)
}
\arguments{
\item{annotations}{A list of Annotation objects.}

\item{domain}{Domain of the annotations, either "time", "frequency", or "both".}

\item{decreasing}{If TRUE, sort in decreasing order.}
}
\value{
A list of Annotation objects.
}
\description{
Sorts a list of annotations by either start time, frequency, or both.
}
