% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundSpeed_cramer.R
\name{soundSpeed_cramer1993}
\alias{soundSpeed_cramer1993}
\title{Get the speed of sound in air using Cramer (1993)}
\usage{
soundSpeed_cramer1993(temp, temp.unit = "C", pressure,
  pressure.unit = "kPa", RH, MoleFracCO2 = 400^-6)
}
\arguments{
\item{temp}{Temperature}

\item{temp.unit}{Temperature unit}

\item{pressure}{Pressure}

\item{pressure.unit}{Pressure unit}

\item{RH}{Relatove humidity}

\item{MoleFracCO2}{Mole fraction of CO2}
}
\description{
TODO: Description
}
