% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundSpeed.R
\name{soundSpeedBMD}
\alias{soundSpeedBMD}
\title{Get the speed of sound in a medium using bulk modulus and density}
\usage{
soundSpeedBMD(bm, d)
}
\arguments{
\item{bm}{Bulk modulus}

\item{d}{Density}
}
\description{
TODO: Description
}
