% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundSpeed.R
\name{soundSpeedWF}
\alias{soundSpeedWF}
\title{Get the speed of sound in a medium using wavelength and frequency}
\usage{
soundSpeedWF(wl, f)
}
\arguments{
\item{wl}{Wavelength}

\item{f}{Frequency}
}
\description{
TODO: Description
}
