% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specStats.R
\name{specStats}
\alias{specStats}
\title{Calculate and plot statistics on a  frequency spectrum}
\usage{
specStats(spectra, stats = "minMax", line.col = "black", ribbon.col = "grey70")
}
\arguments{
\item{spectra}{A list of spectra}

\item{stats}{Either minMax or sd}

\item{line.col}{Colour for the line}

\item{ribbon.col}{Colour for the ribbon}
}
\value{
A ggplot2 object
}
\description{
Given a list of outputs from meanspec generates a plot with the mean shown by
a line, and either the minimum/maximum values or one standard deviation
shown by a ribbon.
}
