% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scikit-maad-classes.R
\docType{class}
\name{spectrogram_maad-class}
\alias{spectrogram_maad-class}
\title{An S4 class to represent a spectrogram from scikit-maad.}
\description{
An S4 class to represent a spectrogram from scikit-maad.
}
\section{Slots}{

\describe{
\item{\code{Sxx}}{The spectrogram matrix}

\item{\code{tn}}{The time vector}

\item{\code{fn}}{The frequency vector}

\item{\code{extents}}{The extents of the spectrogram. A list of tmin, tmax, fmin, fmax.}
}}

