% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STP.R
\docType{data}
\name{STP}
\alias{STP}
\title{STP: Standard Temperature and Pressure}
\format{An object of class \code{list} of length 2.}
\usage{
STP
}
\description{
Dataset compiled from various sources for differing values of STP.
}
\keyword{datasets}
