% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaggedWave.R
\docType{class}
\name{TaggedWave-class}
\alias{TaggedWave-class}
\title{A S4 class for tagged waves}
\description{
The TaggedWave class extended the Wave class from the tuneR package so
that it can include extended metadata and the results of analyses.
}
\section{Slots}{

\describe{
\item{\code{metadata}}{A list for storing metdata.}

\item{\code{analyses}}{A list for storing analyses.}
}}

