% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaggedWave.R
\name{tagSlots}
\alias{tagSlots}
\title{List of extra slots for TaggedWave and TaggedWaveMC}
\usage{
tagSlots()
}
\description{
This function is used to provide a list of new slots for use when extending Wave
or WaveMC objects to TaggedWave or TaggedWaveMC.
}
