% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaggedWave.R
\name{tagWave}
\alias{tagWave}
\title{Tag a Wave or WaveMC object}
\usage{
tagWave(w)
}
\arguments{
\item{w}{A Wave or WaveMC object.}
}
\value{
A TaggedWave or TaggedWaveMC object.
}
\description{
This function takes a Wave/WaveMC object and returns a corresponding
TaggedWave/TaggedWaveMC object.
}
