% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeRegion.R
\docType{class}
\name{TimeRegion-class}
\alias{TimeRegion-class}
\title{An S4 class to represent a TimeRegion within a Wave object.}
\description{
An S4 class to represent a TimeRegion within a Wave object.
}
\section{Slots}{

\describe{
\item{\code{from}}{Start position}

\item{\code{to}}{End position}

\item{\code{unit}}{Time unit (one of seconds, minutes, hours)}
}}

