% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseudoWave.R
\docType{methods}
\name{*,numeric,PseudoWave-method}
\alias{*,numeric,PseudoWave-method}
\title{Numeric multiplication by PseudoWave}
\usage{
\S4method{*}{numeric,PseudoWave}(e1, e2)
}
\arguments{
\item{e1}{Input 1}

\item{e2}{Input 2}
}
\value{
A PseudoWave object with its scale multiplied by e1.
}
\description{
Numeric multiplication by PseudoWave
}
