% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dielPlot.R
\name{tzRot}
\alias{tzRot}
\title{Converts a timezone offset into a rotation}
\usage{
tzRot(tz, init = pi)
}
\arguments{
\item{tz}{Timezone numeric}

\item{init}{Initial rotation. Defaults to pi.}
}
\value{
The rotation of the plot, in radians.
}
\description{
Given a timezone offset in hours returns a rotation in radians to apply to values for a diel plot.
}
