% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaggedWave.R
\name{untagWave}
\alias{untagWave}
\title{Untag a TaggedWave or TaggedWaveMC object}
\usage{
untagWave(w)
}
\arguments{
\item{w}{A TaggedWave or TaggedWaveMC object.}
}
\value{
A Wave or WaveMC object.
}
\description{
This function takes a TaggedWave/TaggedWaveMC object and returns a corresponding
Wave/WaveMC object.
}
\examples{
\dontrun{
w <- noise("white")
tw <- tagWave(w)
w2 <- untagWave(tw)
}
}
