% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reticulate.R
\name{r_to_py.Wave}
\alias{r_to_py.Wave}
\alias{py_to_r.Wave}
\title{Wave to python conversion}
\usage{
r_to_py.Wave(rx, convert = FALSE)

py_to_r.Wave(px, convert = FALSE)
}
\arguments{
\item{rx}{A Wave object.}

\item{convert}{Logical indicating whether to convert the Wave object to an R object.}

\item{px}{A Python representation of a wave.}
}
\description{
Helper functions for reticulate to move Wave objects between R and Python
}
