% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaveFilter.R
\docType{class}
\name{WaveFilter-class}
\alias{WaveFilter-class}
\title{A S4 class to represent a filter for a Wave object}
\description{
A WaveFilter object is an object containing information necessary for the
filterw function to apply the filter to a Wave object. This is designed to
allow a pipe operator (either magrittr or base R) to be used to apply filters
to a Wave in a pipeline.
}
\section{Slots}{

\describe{
\item{\code{module}}{Module the filter function is found in}

\item{\code{func}}{Name of function}

\item{\code{params}}{List of additional parameters to pass to the function}
}}

