% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelength.R
\name{wavelength}
\alias{wavelength}
\title{Get the wavelength from frequency and speed of sound}
\usage{
wavelength(frequency, speed = soundSpeed(medium = "air"), unit = "m")
}
\arguments{
\item{frequency}{Frequency of the sound in Hertz}

\item{speed}{Speed of sound (defaults to the speed of sound in air)}

\item{unit}{Unit to return the wavelength in, either "m" or "cm"}
}
\value{
Wavelength of the sound in the unit specified
}
\description{
Calculates the wavelength of a sound wave given the frequency and speed of sound in that medium.
}
\examples{
wl <- wavelength(1000)
wl <- wavelength(1000, speed=soundSpeed(medium="steel"))
wl <- wavelength(1000, unit="cm")
}
