% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowing.R
\name{windowing.functions}
\alias{windowing.functions}
\title{List available windowing functions}
\usage{
windowing.functions()
}
\description{
Lists all available windowing functions.
}
\examples{
\dontrun{
windowing.functions()
}
}
\keyword{wave}
