% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\name{writeAnnotationWave}
\alias{writeAnnotationWave}
\title{Save a time region defined by an Annotation object as a Wave file}
\usage{
writeAnnotationWave(annotation, wave = NULL)
}
\arguments{
\item{annotation}{An Annotation object.}

\item{wave}{(Optional) A Wave object, if not given will load the wave file from Annotation filename.}
}
\value{
No return value, called for its side effect of writing a Wave file.
}
\description{
Save a time region defined by an Annotation object as a Wave file
}
