% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audacity.R
\name{writeAudacityLabels}
\alias{writeAudacityLabels}
\title{Write an Audacity label file}
\usage{
writeAudacityLabels(annotations, file)
}
\arguments{
\item{annotations}{A list of \code{Annotation} objects.}

\item{file}{Path to the Audacity label file.}
}
\value{
No return value, called for its side effect of writing a label file.
}
\description{
Writes a list of \code{Annotation} objects to an Audacity label file.
\cr\cr
Internally this uses the \code{write.audacity()} function from the \code{seewave}
package \insertCite{seewave2008}{sonicscrewdriver}.
}
\references{
\insertAllCited{}
}
