% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearlyPlot.R
\name{yearlyFraction}
\alias{yearlyFraction}
\title{Calculate the fraction of a year given by a value}
\usage{
yearlyFraction(t, year = 2022, input = "POSIX", unit = "radians")
}
\arguments{
\item{t}{Object to be converted to a fraction}

\item{year}{Year to calculate fractions of (allows for leap years)}

\item{input}{One of POSIXlt (default)}

\item{unit}{If set to radians outputs a position around a circle. If set to fraction outputs the raw fraction.}
}
\value{
The position of the date within the year, in radians or as a fraction of a year.
}
\description{
Given an object that can be coerced to POSIXlt, return the fraction of a year represented by the object.
}
