% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearlyPlot.R
\name{yearlyLabels}
\alias{yearlyLabels}
\title{Generate labels for a yearly plot}
\usage{
yearlyLabels()
}
\value{
A character vector of twelve month abbreviations.
}
\description{
Generates monthly labels for a yearlyPlot().
}
