% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zerocrossing.R
\name{zerocross}
\alias{zerocross}
\title{Identify zero crossings in a Wave object}
\usage{
zerocross(wave)
}
\arguments{
\item{wave}{A Wave object}
}
\value{
A vector of zero crossing locations
}
\description{
Returns a vector of the position (in samples) of zero crossings in
a Wave object
}
\examples{
\dontrun{
zerocross(sheep)
}
}
