% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectraMaths.R
\name{zeroSpectrum}
\alias{zeroSpectrum}
\title{Zero spectrum}
\usage{
zeroSpectrum(s1)
}
\arguments{
\item{s1}{Spectrum to emulate the structure of.}
}
\value{
A zero-valued spectrum.
}
\description{
This function takes a spectrum from seewave and creates a new zero-valued spectrum with the same structure.
}
\examples{
\dontrun{
zeroSpectrum(spec)
}

}
