% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allChannels.R
\name{allChannels}
\alias{allChannels}
\title{Apply a function to all channels of a Wave or WaveMC object}
\usage{
allChannels(w, FUN, cl = NULL, channel.param = "channel", ...)
}
\arguments{
\item{w}{A Wave or WaveMC object}

\item{FUN}{Function to apply to the wave.}

\item{cl}{Optionally a cluster for parallel calculation.}

\item{channel.param}{Optional. Name of the channel paramater to FUN.}

\item{...}{Optional. Additional paramaters to pass to FUN.}
}
\value{
A list of outputs.
}
\description{
Some functions (e.g. ffilter from seewave) only operate on a single
channel at a time. This function applies the function to each channel
and returns a list of analyses.
}
