% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audioblast.R
\name{audioblast}
\alias{audioblast}
\title{Get data or analyses from audioBlast}
\usage{
audioblast(type, name, endpoint = NULL, check = TRUE, ...)
}
\arguments{
\item{type}{One of data, analysis, standalone.}

\item{name}{Name of data or analysis source.}

\item{endpoint}{Optionally specify endpoint of an audioBlast module.}

\item{check}{Logical. Performs sanity check on input before sending to audioBLAST.}

\item{...}{Fields and values to filter on. Any field defined by audioBLAST.}
}
\value{
A data frame
}
\description{
Search for recordings or analyses on audioBlast.
}
\examples{
\dontrun{
audioblast("data", "recordings", taxon="Gryllotalpa vineae")
}

}
