% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audiomoth.R
\name{audiomoth_config}
\alias{audiomoth_config}
\title{Read AudioMoth configuration file}
\usage{
audiomoth_config(filename)
}
\arguments{
\item{filename}{Path to the configuration file to read}
}
\value{
A data frame of matching annotations
}
\description{
Reads and parses an AudioMoth configuration file.
}
\examples{
\dontrun{
audiomoth_config("./CONFIG.TXT")
}

}
