% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoBandPass.R
\name{autoBandPass}
\alias{autoBandPass}
\title{AUtoamtic band pass}
\usage{
autoBandPass(wave, sd, nsd)
}
\arguments{
\item{wave}{A Wave object}

\item{sd}{Standard deviation of filter}

\item{nsd}{Number of stanbard deviations eahc side of the mean to pass}

\item{...}{Additional parameters to provide to the filter function}
}
\description{
Allows for automatic complex frequency filtering of Wave objects, using a normal distribution.
}
