% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allChannels.R
\name{channels_se}
\alias{channels_se}
\title{Helper function for processing channel data in bioacoustic indices}
\usage{
channels_se(...)
}
\arguments{
\item{...}{Export form a bioaocustoc index function.}
}
\description{
Used to provide allChannels with relevant parameters.
}
