% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFileSizes.R
\name{convert2bytes}
\alias{convert2bytes}
\title{Convert bits to bytes}
\usage{
convert2bytes(S, input = "bits")
}
\arguments{
\item{S}{The value to convert}

\item{input}{The unit to convert, allowed values are "bits", "bytes", "kB", "MB", "GB"}
}
\value{
The numeric value in bytes
}
\description{
Converts file size measurements into bytes.
}
