% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTemperatures.R
\name{convert2Celsius}
\alias{convert2Celsius}
\title{Convert temperature to Celsius}
\usage{
convert2Celsius(temp, input = "K")
}
\arguments{
\item{temp}{The value of the temperature to convert}

\item{input}{The unit of the temperature to convert, allowed values are "K", "F".}
}
\value{
Numeric value in degrees Celsius
}
\description{
Converts temperature measurements into Celsius
}
\examples{
convert2Celsius(15, input="K")
convert2Celsius(15, input="F")

}
