% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPressures.R
\name{convert2dyne_cm2}
\alias{convert2dyne_cm2}
\title{Convert pressure to dyne per square centimetre}
\usage{
convert2dyne_cm2(P, input = "kPa")
}
\arguments{
\item{P}{The value of the pressure to convert}

\item{input}{The unit of the pressure to convert, allowed values are "kPa", "P".}
}
\value{
The numeric value in dyne per square centimetre.
}
\description{
Converts pressure measurements into dyne per square centimetre
}
\examples{
convert2dyne_cm2(1, input="Pa")
convert2dyne_cm2(1, input="kPa")
}
