% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTemperatures.R
\name{convert2Fahrenheit}
\alias{convert2Fahrenheit}
\title{Convert temperature to Farenheit}
\usage{
convert2Fahrenheit(temp, input)
}
\arguments{
\item{temp}{The value of the temperature to convert}

\item{input}{The unit of the temperature to convert, allowed values are "K", "C".}
}
\description{
Converts tmperature measurements into Farenheit
}
