% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corWaveMC.R
\name{corWaveMC}
\alias{corWaveMC}
\title{Correlate channels in a WaveMC object}
\usage{
corWaveMC(wave, times, window, temp = 25, cluster = NULL)
}
\arguments{
\item{wave}{A WaveMC object}

\item{times}{One or more times of events to correlate}

\item{window}{Width of the window to correlate in seconds (centred on times)}

\item{temp}{Air temperature in Celsius}

\item{cluster}{A cluster for parallel execution}
}
\value{
List of corenv lists for events, and a list of the time differences between channels
}
\description{
Uses the corenv function from seewave to calculate the envelope correlation for timed
events between the channels of a WaveMC object
}
