% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutws.R
\name{cutws}
\alias{cutws}
\title{Cut wave by samples}
\usage{
cutws(wave, from, to, plot = FALSE)
}
\arguments{
\item{wave}{A Wave object}

\item{from}{First sample to return}

\item{end}{Last sample to return}
}
\description{
Extract a section of a Wave object based on sample positions
}
