% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutws.R
\name{cutws}
\alias{cutws}
\title{Cut wave by samples}
\usage{
cutws(wave, from, to, plot = FALSE)
}
\arguments{
\item{wave}{A Wave object}

\item{from}{First sample to return}

\item{to}{Last sample to return}

\item{plot}{If TRUE shows the cut region within the original waveform}
}
\value{
A Wave object
}
\description{
Extract a section of a Wave object based on sample positions
}
\examples{
\dontrun{
cutws(sheep, 1, 20)
cutws(sheep, 1, 20, plot=TRUE)
}

}
