% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2Wave.R
\name{data2Wave}
\alias{data2Wave}
\title{Convert data into a Wave object}
\usage{
data2Wave(
  left,
  samp.rate = 44100,
  bit = 16,
  remove.offset = TRUE,
  normalise = TRUE
)
}
\arguments{
\item{left}{Data for  audio channel}

\item{samp.rate}{Sampling rate for Wave object}

\item{bit}{Bit depth of Wave object}

\item{remove.offset}{If TRUE any DC offset is removed}

\item{normalise}{IF TRUE the output Wave is normalised using tuneR}
}
\value{
A mono Wave object.
}
\description{
Make a sequence of data into a normalised Wave object.
}
\examples{
pattern <- seq(from=-1, to=1, length.out=100)
data <- rep.int(pattern, 100)
w <- data2Wave(data)
}
