% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dayPhase.R
\name{dayPhases}
\alias{dayPhases}
\title{Phases of day}
\usage{
dayPhases(time, lat, lon, tz)
}
\arguments{
\item{time}{A time object representing the start time of a recording}

\item{lat}{Latitude of recording device}

\item{lon}{Longitude of recording device}

\item{tz}{Timezone of recording device when recording was made}
}
\description{
Wrapper for suncalc getSunlightTimes that formats output for this package.
}
