% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultCluster.R
\name{defaultCluster}
\alias{defaultCluster}
\title{Create Default Cluster for Windowing}
\usage{
defaultCluster(fork = TRUE)
}
\arguments{
\item{fork}{If TRUE uses forking to create the cluster}
}
\value{
A cluster object for parallel processing
}
\description{
Creates a default cluster using one less than the total cores available on the system. By default this uses forking, which may not be available on 'Windows'.
}
\examples{
\dontrun{
cl <- defaultCluster()
stopCluster(cl)
cl <- defaultCluster(FALSE)
stopCluster(cl)
}

}
\keyword{wave}
