% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dielPlot.R
\name{dielLabels}
\alias{dielLabels}
\title{Generate labels for a diel plot}
\usage{
dielLabels(format = "clock24")
}
\arguments{
\item{format}{One of clock24 (default) or clock12}
}
\description{
Generates labels for a dielPlot() in 12- or 24-hour format. Labels are generated
at three hourly intervals.
}
\examples{
dielLabels()
dielLabels("clock12")
}
