% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dielPlot.R
\name{dielPlot}
\alias{dielPlot}
\title{Create a diel plot}
\usage{
dielPlot(
  date,
  lat,
  lon,
  limits = c(0, 2),
  plot = NULL,
  rot = tzRot(0),
  method = "plotrix",
  legend = F
)
}
\arguments{
\item{date}{Date to plot.}

\item{lat}{Numeric latitude.}

\item{lon}{Numeric longitude.}

\item{limits}{Plotting limits of the daylight regions, default to c(1,2)}

\item{plot}{Character vector of components to plot}

\item{rot}{Either "Solar Noon" or an offset calculated by tz}

\item{method}{Plotting library to use}

\item{legend}{Whether to show a legend}
}
\value{
Called for its side effect of drawing a plot. The return value is that of the underlying plotting function and should not be relied on.
}
\description{
A diel plot shows the times of night, twilight and the maximum altitude of the sun for a given date.
}
