% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dielPlot.R
\name{dielPlot}
\alias{dielPlot}
\title{Create a diel plot}
\usage{
dielPlot(
  date,
  lat,
  lon,
  limits = c(0, 2),
  plot = NULL,
  method = "plotrix",
  legend = F
)
}
\arguments{
\item{date}{Date to plot.}

\item{lat}{Name of data or analysis source.}

\item{lon}{Optionally specify endpoint of an audioBlast module.}

\item{plot}{Logical. Performs sanity check on input before sending to audioBLAST.}
}
\description{
A diel plot shows the times of night, twilight and the maximum altitude of the sun for a given date.
}
