% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dielPlot.R
\name{dielPositions}
\alias{dielPositions}
\title{Generate positions of labels for a diel plot}
\usage{
dielPositions(format = "3hourly")
}
\arguments{
\item{format}{One of "3hourly" (default), "hours", or "minutes"}
}
\value{
A numeric vector of label positions, in radians.
}
\description{
Generates positions for three-hourly labels of a dielPlot() in radians.
}
\examples{
dielPositions()
dielPositions("hours")
dielPositions("minutes")
}
