% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dielPlot.R
\name{dielRings}
\alias{dielRings}
\title{Plot rings on a diel plot}
\usage{
dielRings(
  names,
  starts,
  ends,
  cols = "grey",
  format = "HHMM",
  limits = c(1, 2),
  legend = T
)
}
\arguments{
\item{names}{Labels for the rings}

\item{starts}{Start times for rings in HHMM string format}

\item{ends}{End times for rings in HHMM string format}

\item{cols}{Colours of the rings}

\item{format}{Defaults to HHMM}

\item{limits}{Region of a dielPlot() to plot rings. Defaults to c(1,2)}

\item{legend}{Boolean. Whether to plot a legend.}
}
\description{
Plot rings on a diel plot.
}
