% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dolbear.R
\name{dolbear}
\alias{dolbear}
\title{Dolbear's law}
\usage{
dolbear(n = NULL, t = NULL, species = "Oecanthus fultoni")
}
\arguments{
\item{n}{Chirps per minute}

\item{t}{Temperature in Celsius}

\item{species}{Species to use (by default Oecanthus fultoni)}
}
\value{
Missing value of n or t
}
\description{
Calculates either chirps per minute based on temperature or vice versa
using Dolbear's law (or equivalent laws for other species)
}
\examples{
dolbear(n=6)
dolbear(t=25)

}
