% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dutyCycle.R
\name{dutyCycle}
\alias{dutyCycle}
\title{Calculate the duty cycle of a wave}
\usage{
dutyCycle(wave, limit = 0.1, output = "unit")
}
\arguments{
\item{wave}{A Wave object}

\item{limit}{Threshold above which to consider the signal}

\item{output}{If "unit" the duty cycle will be in the range 0-1. For a percntage use "percent".}
}
\description{
Proportion of a wave with signal above the limit
}
