% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearlyPlot.R
\name{emptyYearly}
\alias{emptyYearly}
\title{Create an empty diel plot}
\usage{
emptyYearly(year = 2022, rot = pi)
}
\arguments{
\item{rot}{Rotation of the origin (defaults to pi)}

\item{method}{Plotting package to use}
}
\description{
Create a yearly plot with labels but without sun or night duration plotted.
}
