% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropyStats.R
\name{entropyStats}
\alias{entropyStats}
\title{Various measurements of frequency values for a Wave object}
\usage{
entropyStats(wave)
}
\arguments{
\item{wave}{A Wave object}
}
\value{
A list of spectral entropy types.
}
\description{
Calculates the peak, centre, bandwidth and quality factor. The quality factor (Q) is calculated at both
-3dB and -10dB as discussed by Bennett-Clark (1999)  \url{doi:10.1080/09524622.1999.9753408}.
}
\examples{
\dontrun{
entropyStats(sheep)
}

}
