% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoBandPass.R
\name{freqfilter}
\alias{freqfilter}
\title{Frequency Filtering}
\usage{
freqfilter(wave, FUN = NULL, highpass = NULL, ...)
}
\arguments{
\item{wave}{A Wave object}

\item{FUN}{A function to provide the filter value at a frequency}

\item{highpass}{If numeric cut off frequewncies below this cut off}

\item{...}{Additional parameters to provide to the filter function}
}
\description{
Allows for complex frequency filtering of Wave objects.
}
