% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoBandPass.R
\name{freqfilter}
\alias{freqfilter}
\title{Frequency Filtering}
\usage{
freqfilter(wave, FUN = NULL, ...)
}
\arguments{
\item{wave}{A Wave object}

\item{FUN}{A function to provide the filter value at a frequency}

\item{...}{Additional parameters to provide to the filter function}
}
\description{
Allows for complex frequency filtering of Wave objects.
}
\examples{
library(tuneR)
wave <- readWave(system.file("extdata", "1.wav", package="tdsc"))
t <- tdsc(wave)
emptyBands(t,t)

}
\references{
Stammers (2011) “Audio Event Classification for Urban Soundscape Analysis”. PhD thesis. University of York.
}
\keyword{TDSC}
