% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{frequencySound}
\alias{frequencySound}
\title{Get the frequency from wavelength and speed of sound}
\usage{
frequencySound(wl, s = soundSpeedMedium("air"))
}
\arguments{
\item{wl}{Wavelength}

\item{s}{Speed of sound (defaults to the speed of sound in air)}
}
\value{
Frequency of the sound in Hertz
}
\description{
Calculates the frequency of a sound wave given the wavelength and speed of sound in that medium.
}
\examples{
f <- frequencySound(wl=100, s=343)
}
