% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencyStats.R
\name{frequencyStats}
\alias{frequencyStats}
\title{Various measurements of frequency values for a Wave object}
\usage{
frequencyStats(wave, warn = TRUE, plot = FALSE)
}
\arguments{
\item{wave}{A Wave object}

\item{warn}{If TRUE provides warnings when values are not consistent}

\item{plot}{IF TRUE displays values}
}
\description{
Calculates the peak, centre, bandwidth and quality factor. The quality factor (Q) is calculated at both
-3dB and -10dB as discussed by Bennett-Clark (1999)  <doi: 10.1080/09524622.1999.9753408>.
}
