% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateNoise.R
\name{generateNoise}
\alias{generateNoise}
\title{Add noise to a soundwave}
\usage{
generateNoise(wave, noise = c("white"), noiseAdd = FALSE,
  noiseRatio = 0.5, output = "file", plot = FALSE)
}
\arguments{
\item{wave}{Wave file to add noise to}

\item{noise}{Vector of noise to add (unif, gaussian, white, pink, power, red, frequency of a sine wave in Hz, or filename)}

\item{noiseAdd}{If TRUE all noise sources are added to wave. If FALSE separate outputs are created for each noise source.}

\item{noiseRatio}{Ratio of maximum noise amplitude to the maximum amplitude in wave}

\item{output}{TODO: Is this implemented?}

\item{plot}{If TRUE various plots are made to show how noise is added.}
}
\value{
A list of Wave objects with the required noise added.
}
\description{
Adding noise to a soundwave allows for testing of the robustness of automated identification algorithms
to noise.
}
