% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateTimeMask.R
\name{generateTimeMask}
\alias{generateTimeMask}
\title{Generate time masked Wave-like objects}
\usage{
generateTimeMask(wave, method = "squarewave", dutyCycle = 0.95, n.periods = 10)
}
\arguments{
\item{wave}{A \code{Wave}-like object (or a list of \code{Wave}-like objects).}

\item{method}{The method to use for time masking (one of "squarewave", "random).}

\item{dutyCycle}{The duty cycle of the output. A value of 0.95 means that 5\%
of the time is masked.}

\item{n.periods}{The number of waves to generate in the squarewave method.}
}
\description{
Given a \code{Wave}-like object (or a list of \code{Wave}-like objects), generate
new \code{Wave}-like objects with time masking.
}
