% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFileSizes.R
\name{humanBytes}
\alias{humanBytes}
\title{Converts bytes in human readable form}
\usage{
humanBytes(S)
}
\arguments{
\item{S}{Number of bytes}
}
\value{
String in human readable format
}
\description{
Given an input of bytes calculates the result in a sensible output unit (e.g.
MB, GB, PB).
}
