% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTime.R
\name{humanTime}
\alias{humanTime}
\title{Converts seconds in human readable form}
\usage{
humanTime(S)
}
\arguments{
\item{S}{Number of seconds}
}
\value{
String in human readable format
}
\description{
Given an input of bytes calculates the result in a sensible output unit (e.g.
minutes, hours).
}
