% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTime.R
\name{humanTime}
\alias{humanTime}
\title{Converts time to human readable form}
\usage{
humanTime(S, unit = "seconds")
}
\arguments{
\item{S}{Time to convert in unit}

\item{unit}{The unit of time to convert}
}
\value{
String in human readable format
}
\description{
Given an input of bytes calculates the result in a sensible output unit (e.g.
minutes, hours).
}
