% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jitter.R
\name{jitter}
\alias{jitter}
\title{Calculate the jitter in a Wave object}
\usage{
jitter(wave, method = "absolute")
}
\arguments{
\item{wave}{A Wave object}

\item{method}{One of "absolute" or "relative"}
}
\value{
A vector of zero crossing locations
}
\description{
Jitter is a measure of the variability of periods in the waveform. Relative
jitter is scaled by the jitter in the analysed waveform.
}
\examples{
\dontrun{
jitter(sheep, method="absolute")
jitter(sheep, method="relative")
}
}
