% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{labelReduction}
\alias{labelReduction}
\title{Combines labels which overlap into single continuous regions}
\usage{
labelReduction(t)
}
\arguments{
\item{t}{Transcript from Google Speech API}
}
\value{
A list containing start and end times of speech containing regions
}
\description{
Takes labels from Google Speech API transcript and combines overlapping labels.
}
\examples{
\dontrun{
labelReduction(t)
}

}
