% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scikit-maad-indices.R
\name{maad_number_of_peaks}
\alias{maad_number_of_peaks}
\title{Compute the number of peaks using scikit-maad}
\usage{
maad_number_of_peaks(
  object,
  mode = "dB",
  min_peak_val = NULL,
  min_freq_dist = 200,
  slopes = c(1, 1),
  prominence = 0,
  maad = NULL
)
}
\arguments{
\item{object}{A Wave object or a spectrogram_maad object generated by
\code{\link{maad_spectrogram}}. If a Wave-like object is provided, the
spectrogram will be calculated using the default parameters.}

\item{mode}{One of "dB", "linear".}

\item{min_peak_val}{Only peaks above this threshold will be considered.}

\item{min_freq_dist}{Frequency threshold parameter (in Hz). If the frequency
difference of two successive peaks is less than this threshold, then the
peak of highest amplitude will be kept only.}

\item{slopes}{Refers to the amplitude slopes of the peak. The first value is
the left slope and the second value is the right slope. Only peaks with
higher slopes than threshold values will be kept.}

\item{prominence}{The first element is the minimal prominence and the second
element is the maximal prominence. If a single number is provided it is
interpreted as the minimal value, and no maximial value will be used.}

\item{maad}{An optional maad object. If not provided, one will be created
using \code{getMaad()}.}
}
\value{
Numeric number of peaks
}
\description{
Count the number of frequency peaks on a mean spectrum. This function was
adapted from the function fpeaks of the R package Seewave
\insertCite{seewave2008}{sonicscrewdriver}.
}
\details{
For addition documentation see
\url{https://scikit-maad.github.io/generated/maad.features.number_of_peaks.html}
}
