% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcis.R
\name{map2RGB}
\alias{map2RGB}
\title{Map three vectors to RGB}
\usage{
map2RGB(red, green, blue)
}
\arguments{
\item{red}{The red channel vector}

\item{green}{The green channel vector}

\item{blue}{The blue channel vector}
}
\value{
A vector of RGB values
}
\description{
Maps three vectors of equal length to RGB for use in false-colour index
spectrograms
}
