% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntd.R
\name{ntd}
\alias{ntd}
\title{Natural Time Domain}
\usage{
ntd(wave, events, FUN, normalise = FALSE, argument = "wave", ...)
}
\arguments{
\item{wave}{A Wave object containing pulses}

\item{events}{Onset of detected events, e.g. from pulseDetection()}

\item{FUN}{The function to run}

\item{normalise}{If TRUE the output is a probability density}

\item{argument}{If "wave" supplies a weave object to the function, if "vector" supplies the left channel as a numeric vector.}

\item{...}{Additional arguments to FUN}
}
\value{
A list of outputs form the applied function
}
\description{
Runs a function on the wave and outputs values in the Natural Time Domain (see Varotsos, Sarlis & Skordas(2011) \url{doi:10.1007/978-3-642-16449-1}).
}
