% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseFilename.R
\name{parseFilename}
\alias{parseFilename}
\title{Parse a filename}
\usage{
parseFilename(string, format = NULL, timezone = NULL)
}
\arguments{
\item{string}{A filename.}

\item{format}{Optionally force a given format - "timestamp".}

\item{timezone}{Optionally set a timezone.}
}
\value{
A list of string, type of match, datetime.
}
\description{
Attempts to extract meaningful information from a filename.
}
\examples{
parseFilename("5E90A4D4.wav")

}
