% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseFilename.R
\name{parseFilename}
\alias{parseFilename}
\title{Parse a filename}
\usage{
parseFilename(string)
}
\arguments{
\item{string}{A filename}
}
\value{
A list of raw results, plus calculated values for date, time and device.
}
\description{
Attempts to extract meaningful information from a filename.
}
