% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PseudoWave.R
\docType{class}
\name{PseudoWave-class}
\alias{PseudoWave-class}
\title{An S4 class to represent a PseudoWave object that is converted to a
Wave object when operated on.}
\description{
An S4 class to represent a PseudoWave object that is converted to a
Wave object when operated on.
}
\section{Slots}{

\describe{
\item{\code{type}}{Type of PseudoWave (e.g. "noise")}

\item{\code{subtype}}{Subtype of PseudoWave (e.g. "white" if type is "noise")}

\item{\code{scale}}{The Wave channels are multiplied by this value}

\item{\code{offset}}{This value is added to the  Wave channels}

\item{\code{seed}}{Random seed for reproducible output, NA for no seed}

\item{\code{scale}}{Logical. Whether to use the random seed value}

\item{\code{params}}{List of additional parameters to pass to generating function}
}}

