% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ste.R
\name{pulseDetection}
\alias{pulseDetection}
\title{Pulse detection}
\usage{
pulseDetection(wave, method = "dietrich2004", ...)
}
\arguments{
\item{wave}{A Wave object}

\item{method}{Which method to use}

\item{...}{Other arguments to pass to pulse detection function}
}
\description{
Detects pulses in a Wave.
}
