% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulseDetection.R
\name{pulseDetection}
\alias{pulseDetection}
\title{Pulse detection}
\usage{
pulseDetection(wave, method = "simple", ...)
}
\arguments{
\item{wave}{A Wave object containing pulses}

\item{method}{Which method to use for pulse detection}

\item{...}{Other arguments to pass to pulse detection function}
}
\description{
Detects pulses in a Wave, defaults to using Dietrich (2004).
}
