% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reticulate-virtualenv.R
\name{pythonInstall}
\alias{pythonInstall}
\title{Install sonicscrewdriver Python environment}
\usage{
pythonInstall(unattended = FALSE)
}
\arguments{
\item{unattended}{If TRUE then the function will not prompt the user to install
the environment in a non-interactive session.}
}
\description{
This function installs scikit-maad in the \code{sonicscrewdriver} environment using \code{reticulate}.
}
\examples{
\dontrun{
pythonInstall()
pythonInstall(unattended=TRUE)
}
}
