% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radar.R
\name{radarRange}
\alias{radarRange}
\title{Radar range}
\usage{
radarRange(t, c = soundSpeed(medium = "air"))
}
\arguments{
\item{t}{Time in seconds}

\item{c}{Speed of sound in transmission medium m/s (by default air)}
}
\value{
Distance to object
}
\description{
Calculates the distance of an object based on the round trip time of an
echolocation pulse
}
\examples{
radarRange(2)
radarRange(2, c=343)
radarRange(2, c=soundSpeed(medium = "sea water"))

}
