% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainfallDetection.R
\name{rainfallDetection}
\alias{rainfallDetection}
\title{Rainfal detection}
\usage{
rainfallDetection(wave, method = "bedoya2017", ...)
}
\arguments{
\item{wave}{A Wave object to detect rainfall in}

\item{method}{Which rainfall detection method to use}

\item{...}{Other arguments to pass to pulse detection function}
}
\description{
Detects rainfall in a Wave.
}
