% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainfallDetection.R
\name{rainfallDetection}
\alias{rainfallDetection}
\title{Rainfal detection}
\usage{
rainfallDetection(wave, method = "bedoya2017", ...)
}
\arguments{
\item{wave}{A Wave object to detect rainfall in}

\item{method}{Which rainfall detection method to use ("bedoya2017", "hardRain")}

\item{...}{Other arguments to pass to rain detection function}
}
\description{
Detects rainfall in a Wave. An uncallibrated version of Bedoya (REF) is available in this package. The hardRain
package can also be accessed via this wrapper.
}
