% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAudio.R
\name{readAudio}
\alias{readAudio}
\title{Read an audio file}
\usage{
readAudio(file, mime = "auto", from = 0, to = Inf, units = "seconds")
}
\arguments{
\item{file}{File to read}

\item{mime}{MIME type of file to read, or "auto". Supported types are "audio/x-wav" and "audio/mpeg" (MP3)}

\item{from}{Start point in file to return}

\item{to}{End point in file to return}

\item{units}{One of "samples", "seconds", "minutes", "hours". Default is "seconds".}
}
\value{
A Wave object
}
\description{
This file is used to read an audio file and return a Wave object, it is an abstraction
function for various specific audio reading functions. If no existing method can be identified
an attempt is made to use the av package to read the audio.
}
