% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respeaker6.R
\name{readRespeaker6}
\alias{readRespeaker6}
\title{Read a file from Seeed Studio Respeaker 6 mic array}
\usage{
readRespeaker6(filename, from = 1, to = Inf, units = "samples", header = FALSE)
}
\arguments{
\item{filename}{file to read.}

\item{from}{Where to start reading the wave in units.}

\item{to}{Where to stop reading the wave in units.}

\item{units}{Units in which from and to is given, the default is "samples", but can be set to time intervals such as "seconds".}

\item{header}{If TRUE, just header information of the Wave file are returned, otherwise (the default) the whole Wave object.}
}
\value{
A WaveMC object.
}
\description{
The Seeed Studio Respeaker-6 when used as described in the documentation saves an eight
channel audio file with channels 7 and 8 not containing input audio. This function reads
such a file and saves it as a six channel file.
}
