% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referenceLevels.R
\name{referenceIntensity}
\alias{referenceIntensity}
\title{Reference intensity}
\usage{
referenceIntensity(unit = "watt_cm2")
}
\arguments{
\item{unit}{Unit to return, "watt_cm2"}
}
\description{
Provides the standard reference intensity level.
}
\examples{
ri <- referenceIntensity()
}
