% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referenceLevels.R
\name{referencePressure}
\alias{referencePressure}
\title{Reference pressure}
\usage{
referencePressure(unit = "Pa")
}
\arguments{
\item{unit}{Unit to return, "Pa" or "dyne_cm2"}
}
\description{
Provides the standard reference pressure level.
}
\examples{
rp <- referencePressure()
rp <- referencePressure(unit="dyne_cm2")
}
