% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sDuration.R
\name{sDuration}
\alias{sDuration}
\title{Sample duration}
\usage{
sDuration(n = 1, wave = NULL, samp.rate = NULL)
}
\arguments{
\item{n}{The number of the samples}

\item{wave}{A Wave object containing pulses}

\item{samp.rate}{Integer sampling rate}
}
\description{
Calculates the time represenetd by n samples in a Wave.
}
