% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sDuration.R
\name{sDuration}
\alias{sDuration}
\title{Sample duration}
\usage{
sDuration(n = 1, wave = NULL, samp.rate = NULL)
}
\arguments{
\item{n}{The number of the samples}

\item{wave}{A Wave object containing pulses}

\item{samp.rate}{Integer sampling rate}
}
\value{
A numeirc value in seconds
}
\description{
Calculates the time represented by n samples in a Wave.
}
\examples{
sDuration(n=20, samp.rate=44100)
\dontrun{
sDuration(n=20, wave=sheep)#'
}


}
