% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shimmer.R
\name{shimmer}
\alias{shimmer}
\title{Shimmer calculation}
\usage{
shimmer(wave, method = "dB", pulses = NULL, pulse_method = NULL)
}
\arguments{
\item{wave}{The wave object being analysed.}

\item{method}{One of "db".}

\item{pulses}{Precalculated pulses from the pulseDetection function, or a list of onset and offset times (in samples).}

\item{pulse_method}{Shortcut to run pulseDetection from this function.

The "dB" method calculates the variation between  the maximum absolute amplitude
between successive periods (not the peak-to-peak amplitude).}
}
\description{
Calculation of shimmer (variation of amplitude) between consecutive periods.
}
\examples{
\dontrun{
data(sheep)
shimmer(sheep, pulse_method="threshold")
}

}
