% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shimmer.R
\name{shimmer}
\alias{shimmer}
\title{Calculate the shimmer in a Wave object}
\usage{
shimmer(wave)
}
\arguments{
\item{wave}{A Wave object}
}
\value{
A vector of zero crossing locations
}
\description{
Jitter is a measure of the variability of amplitudes within periods in the waveform. Relative
shimmer is scaled by the shimmer in the analysed waveform.
}
\examples{
\dontrun{
shimmer(sheep)
}
}
