% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundSpeed.R
\name{soundSpeed}
\alias{soundSpeed}
\title{Calculate the speed of sound in a medium}
\usage{
soundSpeed(
  medium = NULL,
  wl = NULL,
  f = NULL,
  bulkModulus = NULL,
  density = NULL
)
}
\arguments{
\item{medium}{Propagation medium (e.g. "air")}

\item{wl}{Wavelength}

\item{f}{Frequency}

\item{bulkModulus}{Bulk modulus}

\item{density}{Density}
}
\description{
Given sufficient parameters (i.e. wavelength and frequency, bulk modulus and density) this
function calculates the speed of sound.
}
