% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundSpeed.R
\name{soundSpeed}
\alias{soundSpeed}
\title{Get the speed of sound in a medium}
\usage{
soundSpeed(medium = "air", method = NULL, temperature = NULL)
}
\arguments{
\item{medium}{Propagation medium (default is "air")}

\item{method}{The method used to calculate speed of sound (default is "vague")}

\item{temperature}{The temperature used to calculate the sound speed}
}
\description{
TODO: Description
}
