% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundSpeed.R
\name{soundSpeed}
\alias{soundSpeed}
\title{Calculate the speed of sound in a medium}
\usage{
soundSpeed(wl = NULL, f = NULL, bulkModulus = NULL, density = NULL)
}
\arguments{
\item{wl}{Wavelength}

\item{f}{Frequency}

\item{bulkModulus}{Bulk modulus}

\item{density}{Density}
}
\description{
Given sufficient parameters (i.e. wavelength and frequency, bulk modulus and density) this
function calculates the speed of sound.
}
