% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ste.R
\name{ste}
\alias{ste}
\title{Short term energy}
\usage{
ste(wave, method = "dietrich2004", ...)
}
\arguments{
\item{wave}{A Wave object}

\item{method}{Which method used to calculate the short term energy,
by default dietrich2004 to use Dietrich (2004) <doi:10.1016/j.patcog.2004.04.004>.}

\item{...}{Other arguments to pass to STE function}
}
\value{
A vector of short term energy values
}
\description{
Computes the short term energy of a Wave.
}
\examples{
\dontrun{
ste(sheep, method="dietrich2004")
}

}
