% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeRegion.R
\name{[,Wave,TimeRegion-method}
\alias{[,Wave,TimeRegion-method}
\title{Allow subsetting a Wave object with a TimeRegion}
\usage{
\S4method{[}{Wave,TimeRegion}(x, i)
}
\arguments{
\item{x}{Wave Object}

\item{i}{TimeRegion object}
}
\description{
Allow subsetting a Wave object with a TimeRegion
}
