% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaggedWave.R
\docType{class}
\name{TaggedWaveMC-class}
\alias{TaggedWaveMC-class}
\title{A S4 class for tagged multi-channel waves}
\description{
The TaggedWaveMC class extended the WaveMC class from the tuneR package so
that it can include extended metadata and the results of analyses.
}
\section{Slots}{

\describe{
\item{\code{metadata}}{A list for storing metdata.}

\item{\code{analyses}}{A list for storing analyses.}
}}

