% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationFunctions.R
\name{validateIsWave}
\alias{validateIsWave}
\title{Check an object is a Wave object}
\usage{
validateIsWave(wave)
}
\arguments{
\item{wave}{Object to test}
}
\description{
Helper function to test that the input is a Wave object. Will create an error if not.
}
