% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reticulate.R
\name{r_to_py.Wave}
\alias{r_to_py.Wave}
\alias{py_to_r.Wave}
\title{Wave to python conversion}
\usage{
r_to_py.Wave(rx, convert = FALSE)

py_to_r.Wave(px, convert = FALSE)
}
\description{
Helper functions for reticulate to move Wave objects between R and Python
}
