% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaveFilter.R
\docType{class}
\name{WaveFilter-class}
\alias{WaveFilter-class}
\title{WaveFilter object for audio filters}
\description{
A \code{WaveFilter} object is an object containing information necessary for the
\code{filterWave()} function to apply the filter to a \code{Wave} or \code{TaggedWave}
object. This is designed to allow a pipe operator (either magrittr or base R)
to be used to apply filters to a Wave in a pipeline. If used with a
\code{TaggedWave} object the function adds information to the \code{processing} slot
documenting its action.
}
\section{Slots}{

\describe{
\item{\code{description}}{Description of the filter.}

\item{\code{func}}{Name of function.}

\item{\code{params}}{List of additional parameters to pass to the function.}
}}

