% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearlyPlot.R
\name{yearlyPlot}
\alias{yearlyPlot}
\title{Create a yearly plot}
\usage{
yearlyPlot(
  year = 2022,
  lat,
  lon,
  limits = c(0, 2),
  plot = NULL,
  method = "plotrix",
  legend = FALSE
)
}
\arguments{
\item{year}{Year to plot (allows for leap years).}

\item{lat}{Numeric latitude.}

\item{lon}{Numeric longitude.}

\item{limits}{Plotting limits of the daylight regions, default to c(1,2)}

\item{plot}{Character vector of components to plot}

\item{method}{Plotting library to use}

\item{legend}{Whether to show a legend}
}
\value{
Called for its side effect of drawing a plot. The return value is that of the underlying plotting function and should not be relied on.
}
\description{
ToDO......
}
