% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearlyPlot.R
\name{yearlyPositions}
\alias{yearlyPositions}
\title{Generate positions of labels for a yearly plot}
\usage{
yearlyPositions(year = 2022, format = "months")
}
\arguments{
\item{year}{Year to calculate}

\item{format}{One of months, mid-months, days}
}
\description{
Generates positions for monthly labels of a dielPlot() in radians. The positions can either be for the
start of the month, or middle of the month.
}
\details{
The function allows for leap years if the year parameter is provided.
}
